+++ task/src/main/java/com/kimtaehoonki/task/comment/presentation/controller/CommentController.java
@@ -1,0 +1,4 @@
+package com.kimtaehoonki.task.comment.presentation.controller;
+
+public class CommentController {
+}
