--- task/src/main/java/com/kimtaehoonki/task/domain/ProjectStatus.java
@@ -1,11 +1,0 @@
-package com.kimtaehoonki.task.domain;
-
-/**
- * 프로젝트 상태 코드.
- * ACTIVATION: 활성, DORMANT: 휴면, EXIT: 종료
- */
-public enum ProjectStatus {
-    ACTIVATION,
-    DORMANT,
-    EXIT
-}
