--- task/src/main/java/com/kimtaehoonki/task/domain/entity/Project.java
@@ -1,34 +1,0 @@
-package com.kimtaehoonki.task.domain.entity;
-
-import com.kimtaehoonki.task.domain.ProjectStatus;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Lob;
-import javax.persistence.Table;
-
-/**
- * Project 도메인객체.
- */
-@Entity
-@Table(name = "projects")
-public class Project {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "project_id")
-    private long id;
-
-    private int adminId;
-
-    private String name;
-
-    @Column(columnDefinition = "TEXT")
-    private String description;
-
-    @Enumerated(EnumType.STRING)
-    private ProjectStatus status;
-}
