--- task/src/main/java/com/kimtaehoonki/task/domain/entity/Tag.java
@@ -1,32 +1,0 @@
-package com.kimtaehoonki.task.domain.entity;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-
-/**
- * Tag 엔티티.
- */
-@Entity
-@Table(name = "tags")
-public class Tag {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "tag_id")
-    private long id;
-
-    @ManyToOne
-    @JoinColumn(name = "project_id")
-    private Project project;
-
-    private String name;
-
-    @ManyToOne
-    @JoinColumn(name = "color_id")
-    private ColorCode colorCode;
-}
