--- task/src/main/java/com/kimtaehoonki/task/presentation/controller/TagController.java
@@ -1,4 +1,0 @@
-package com.kimtaehoonki.task.presentation.controller;
-
-public class TagController {
-}
