--- task/src/main/java/com/kimtaehoonki/task/presentation/dto/CreateProjectRequestDto.java
@@ -1,4 +1,0 @@
-package com.kimtaehoonki.task.presentation.dto;
-
-public class CreateProjectRequestDto {
-}
