--- task/src/main/java/com/kimtaehoonki/task/presentation/dto/RegisterUserResponseDto.java
@@ -1,4 +1,0 @@
-package com.kimtaehoonki.task.presentation.dto;
-
-public class RegisterUserResponseDto {
-}
