+++ task/src/main/java/com/kimtaehoonki/task/project/domain/entity/Project.java
@@ -1,0 +1,33 @@
+package com.kimtaehoonki.task.project.domain.entity;
+
+import com.kimtaehoonki.task.ProjectStatus;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+/**
+ * Project 도메인객체.
+ */
+@Entity
+@Table(name = "projects")
+public class Project {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "project_id")
+    private long id;
+
+    private int adminId;
+
+    private String name;
+
+    @Column(columnDefinition = "TEXT")
+    private String description;
+
+    @Enumerated(EnumType.STRING)
+    private ProjectStatus status;
+}
