+++ task/src/main/java/com/kimtaehoonki/task/project/presentation/dto/UpdateProjectStatusResponseDto.java
@@ -1,0 +1,4 @@
+package com.kimtaehoonki.task.project.presentation.dto;
+
+public class UpdateProjectStatusResponseDto {
+}
