+++ task/src/main/java/com/kimtaehoonki/task/tag/presentation/controller/TagController.java
@@ -1,0 +1,4 @@
+package com.kimtaehoonki.task.tag.presentation.controller;
+
+public class TagController {
+}
