+++ task/src/main/java/com/kimtaehoonki/task/task/presentation/dto/RegisterTaskResponseDto.java
@@ -1,0 +1,4 @@
+package com.kimtaehoonki.task.task.presentation.dto;
+
+public class RegisterTaskResponseDto {
+}
